FIX - Correção de bug: sem verificar user.foto, um erro poderia ser causado caso o usuário não tivesse foto nenhuma salva (null).

+++ src/controllers/user/uploadFoto.js
@@ -11,9 +11,11 @@ module.exports = async (req, res) => {
             const destino = path.join('public/img/user/foto', novoNome);
 
             let user = await getFieldUser(req.iduser, 'idUser, foto');
-            let imgAntiga = path.join('public/img/user/foto', user.foto);
-            if(fs.existsSync(imgAntiga)){
-                fs.unlinkSync(imgAntiga);
+            if(user.foto){
+                let imgAntiga = path.join('public/img/user/foto', user.foto);
+                if(fs.existsSync(imgAntiga)){
+                    fs.unlinkSync(imgAntiga);
+                }
             }
 
             await sharp(tempPath)
